11ty remove prod configuration and use minified resources
This makes for a more consistent developer experience,
which is good, and it means that the site tested locally
is what users will see online. One thing to consider is if
we should have the website-resources loaded from the
server/internet vs locally. The downside of this is that we
must be online to work on the website fully, but honestly
that isn't too unreasonable to do nowadays anyways...

Which of these to use? TODO
{% assign resources = site.resources %}
{% assign resources = "/resources/" %}